+++ Swift Boy/Swift Boy/Emulator/MMU.swift
@@ -211,9 +211,9 @@ public class MMU: MemoryAccessArray {
         }
     }
     
-     public func run(for time: Int16) throws {
+     public func run(for time: UInt8) throws {
         if queue.count > 0 {
-            cycles = cycles + time
+            cycles = cycles + Int16(time)
          
              while cycles > 0 && queue.count > 0 {
                 let cmd = queue.removeFirst()
